pyproject.toml
[project]
name = "sequential-systems"
version = "0.1.0"
description = ""
authors = [
    {name = "Rong Zhou",email = "ron@ronzz.org"}
]
license = {text = "AGPL3.0"}
requires-python = ">=3.10"
dependencies = [
    "pyserial (>=3.5,<4.0)",
    "pyvisa (>=1.16.2,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
